firebaseapp/serializers.py
from rest_framework import serializers

from firebaseapp.models import UserNotification


class UserNotificationSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserNotification
        fields = '__all__'