requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.5.0
anyio==3.7.1
arabic-reshaper==2.1.4
asgi-redis
asgiref
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
azampay==0.4
blinker==1.6.2
boto3==1.26.140
botocore==1.29.140
CacheControl==0.13.1
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.4
cssselect2==0.7.0
daphne==4.0.0
dataclasses-json==0.5.14
distlib==0.3.6
distro==1.8.0
Django==4.1.1
django-cors-headers==3.14.0
django-jsonfield==1.4.1
django-rest-framework==0.1.0
django-storages==1.13.2
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
faiss-cpu==1.7.4
fcm-django==2.0.0
filelock==3.10.0
firebase-admin==6.2.0
Flask==2.3.2
Flask-RESTful==0.3.10
frozenlist==1.4.0
future==0.18.2
google-api-core==2.11.1
google-api-python-client==2.97.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httpx==0.25.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.268
langsmith==0.0.25
lxml==4.9.1
MarkupSafe==2.1.3
marshmallow==3.20.1
msgpack==1.0.5
msgpack-python==0.5.6
mtaa==1.5
multidict==6.0.4
mypy-extensions==1.0.0
mysql==0.0.3
mysqlclient==2.2.0
numexpr==2.8.5
numpy==1.25.2
openai==1.3.4
oscrypto==1.3.0
packaging==23.1
pdf2image==1.16.3
phonenumbers==8.13.26
Pillow==9.2.0
platformdirs==3.1.1
proto-plus==1.22.3
protobuf==4.24.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.8.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
PyPDF2==3.0.1
PyPDF3==1.0.6
pypng==0.20220715.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==5.0.0
regex==2023.8.8
reportlab==3.6.12
requests==2.28.1
rsa==4.9
s3transfer==0.6.1
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
svglib==1.4.1
tenacity==8.2.3
tigopesasdk==0.1.0
tiktoken==0.4.0
tinycss2==1.2.1
tqdm==4.64.1
Twisted==22.10.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.6.2
tzdata==2022.7
tzlocal==4.3
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.13
virtualenv==20.21.0
webencodings==0.5.1
Werkzeug==2.3.6
xhtml2pdf==0.2.8
XlsxWriter==3.1.9
yarl==1.9.2
zope.interface==6.0
